[core] Move RenderData alongside RenderItem
